+++ services/web/server/src/simcore_service_webserver/users/plugin.py
@@ -9,7 +9,7 @@ from servicelib.aiohttp.application_keys import APP_SETTINGS_KEY
 from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setup
 from servicelib.aiohttp.observer import setup_observer_registry
 
-from . import _notifications_handlers, _preferences_handlers, _tokens_rest, _users_rest
+from . import _notifications_rest, _preferences_rest, _tokens_rest, _users_rest
 from ._preferences_models import overwrite_user_preferences_defaults
 
 _logger = logging.getLogger(__name__)
@@ -29,5 +29,5 @@ def setup_users(app: web.Application):
 
     app.router.add_routes(_users_rest.routes)
     app.router.add_routes(_tokens_rest.routes)
-    app.router.add_routes(_notifications_handlers.routes)
-    app.router.add_routes(_preferences_handlers.routes)
+    app.router.add_routes(_notifications_rest.routes)
+    app.router.add_routes(_preferences_rest.routes)
